+++ CHANGELOG.md
@@ -9,3 +9,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Intial release.
+- Support to trigger on hover.
+- Support to animations.